requirements.txt
openai
fastapi
python-dotenv
uvicorn
websockets